switched to repeat-mode and introduced to ExecutionInfo-struct
