+++ simple-counter.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 
-COUNT_TO="$1"
+SLEEP="$1"
+COUNT_TO="$2"
 
 i=1
 while [ $i -le $COUNT_TO ]; do
     echo "counting: $i"
     i=$((i+1))
-    sleep 1
+    sleep $SLEEP
 done
